+++ app/javascript/vue/tasks/digitize/components/shared/depictions.vue
@@ -21,10 +21,18 @@
         :depiction="item"
         edit
       >
-        <div slot="thumbfooter">
+        <div
+          class="horizontal-left-content"
+          slot="thumbfooter">
           <button
             @click="removeDepiction(item)"
-            class="button circle-button btn-delete"/>
+            class="button circle-button btn-delete"
+          />
+          <zoom-image
+            :image-url="getImageDepictionUrl(item)"
+            :width="item.image.width"
+            :height="item.image.height"
+          />
         </div>
       </image-viewer>
     </div>
@@ -38,11 +46,14 @@ import ActionNames from '../../store/actions/actionNames'
 import ImageViewer from 'components/ui/ImageViewer/ImageViewer'
 import Dropzone from 'components/dropzone.vue'
 import { Depiction } from 'routes/endpoints'
+import ZoomImage from './zoomImage.vue'
+import { imageSVGViewBox } from 'helpers/images'
 
 export default {
   components: {
     ImageViewer,
-    Dropzone
+    Dropzone,
+    ZoomImage
   },
   props: {
     actionSave: {
@@ -66,7 +77,7 @@ export default {
       default: 'Drop images or click here to add figures'
     }
   },
-  data: function () {
+  data () {
     return {
       creatingType: false,
       displayBody: true,
@@ -137,6 +148,18 @@ export default {
 
     error (event) {
       TW.workbench.alert.create(`There was an error uploading the image: ${event.xhr.responseText}`, 'error')
+    },
+
+    getImageDepictionUrl (depiction) {
+      const imageWidth = Math.floor(window.innerWidth * 0.75)
+      const imageHeight = (window.innerHeight * 0.40) < 400 ? Math.floor(window.innerHeight * 0.40) : 400
+
+      return imageSVGViewBox(
+        depiction.image.id,
+        depiction.svg_view_box,
+        imageWidth,
+        imageHeight
+      )
     }
   }
 }