+++ CHANGELOG.md
@@ -16,6 +16,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 - Refresh event for smart selectors [#2255]
 - Edit type material in comprehensive form [#2253]
 - Reset selected ids on new search in observation matrices dashboard
+- Tiff images are not render on image viewer
 
 ### Changed
 - Updated Ruby gems